+++ tempodb/backend/gcs/compactor.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"time"
 
 	"cloud.google.com/go/storage"
 	"github.com/google/uuid"
-	"github.com/googleapis/gax-go/v2"
 	"github.com/grafana/tempo/tempodb/backend"
 	"google.golang.org/api/iterator"
 )
@@ -19,12 +17,10 @@ func (rw *readerWriter) MarkBlockCompacted(blockID uuid.UUID, tenantID string) e
 	compactedMetaFilename := backend.CompactedMetaFileName(blockID, tenantID)
 
 	src := rw.bucket.Object(metaFilename)
-	// TODO: Configure retry opts or move it to bucket config.
-	dst := rw.bucket.Object(compactedMetaFilename).Retryer(
-		storage.WithBackoff(gax.Backoff{
-			Initial: time.Second,
-		}),
-		storage.WithPolicy(storage.RetryAlways),
+	dst := rw.bucket.Object(compactedMetaFilename).If(
+		// Only copy if the destination does not exist. Makes the op idempotent.
+		// If the destination exists, we don't care to overwrite it.
+		storage.Conditions{DoesNotExist: true},
 	)
 
 	ctx := context.TODO()