+++ tempodb/backend/gcs/gcs.go
@@ -10,6 +10,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/googleapis/gax-go/v2"
 	"github.com/grafana/tempo/tempodb/backend/instrumentation"
 
 	"cloud.google.com/go/storage"
@@ -270,7 +271,7 @@ func createBucket(ctx context.Context, cfg *Config, hedge bool) (*storage.Bucket
 		instrumentation.PublishHedgedMetrics(stats)
 	}
 
-	// build client
+	// Build client
 	storageClientOptions := []option.ClientOption{
 		option.WithHTTPClient(&http.Client{
 			Transport: transport,
@@ -285,8 +286,17 @@ func createBucket(ctx context.Context, cfg *Config, hedge bool) (*storage.Bucket
 		return nil, errors.Wrap(err, "creating storage client")
 	}
 
-	// build bucket
-	return client.Bucket(cfg.BucketName), nil
+	// Build bucket
+	bucket := client.Bucket(cfg.BucketName)
+	if !hedge {
+		// Retryer is not thread safe, so we only use it when not hedging.
+		bucket = bucket.Retryer(
+			storage.WithBackoff(gax.Backoff{}), // TODO: make this configurable
+			storage.WithPolicy(storage.RetryIdempotent),
+		)
+	}
+
+	return bucket, nil
 }
 
 func readError(err error) error {