+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * [BUGFIX] Apply `rate()` to bytes/s panel in tenant's dashboard. [#2081](https://github.com/grafana/tempo/pull/2081) (@mapno)
 * [BUGFIX] Correctly coalesce trace level data when combining Parquet traces. [#2095](https://github.com/grafana/tempo/pull/2095) (@joe-elliott)
 * [CHANGE] Update Go to 1.20 [#2079](https://github.com/grafana/tempo/pull/2079) (@scalalang2)
+* [BUGFIX] Retry idempotent operations in GCS backend [#2111](https://github.com/grafana/tempo/pull/2111) (@mapno)
 
 ## v2.0.0 / 2023-01-31